+++ History.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * Accept valid `Vary` header string as `field`
   * Add `vary.append` for low-level string manipulation
 
 0.1.0 / 2014-06-05